testing code

+++ src/main/java/org/spring/User.java
@@ -2,10 +2,10 @@ package org.spring;
 
 public class User {
 
-    private long id;
+    private int id;
     private String name;
 
-    public User(long id, String name) {
+    public User(int id, String name) {
         this.id = id;
         this.name = name;
     }
