+++ src/main/java/org/spring/UserController.java
@@ -1,6 +1,6 @@
 package org.spring;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class UserController {
 
-    private final AtomicLong counter = new AtomicLong();
+    private final AtomicInteger counter = new AtomicInteger();
 
     @RequestMapping(value="/user", method=RequestMethod.GET)
     public User user(@RequestParam(value="name", defaultValue="michael") String name) {
